add one case: vscode appears to choose the wrong pathname for WSL
Took me an embarrassingly long time to find this problem: if I drag the file in from Windows File Explorer, I can edit and run the file just fine but vscode's debug output would say it can't find the source file. It was getting the pathname wrong - begingging with C:/Users instead of with /mnt/c/Users as expected. After playing around with launch.json, I discovered it's related to the way I opened the file in the editor, nothing to do with launch.json. If you think this is inappropriate for this doc, please point me to a more appropriate place to report it.